callrules: partition specialized rules from generic ones in preparation for codegen
